restructure(Turbo repo): adding and syncing code across apps
