+++ packages/types/canvas.ts
@@ -1,12 +1,12 @@
 type PropsType = "gradient" | "solid" | "image";
 
 interface ICanvasPropsStyle {
-  id: string;
-  name: string;
-  color?: string;
-  gradient?: string;
-  url?: string;
-  type: PropsType;
+    id: string;
+    name: string;
+    color?: string;
+    gradient?: string;
+    url?: string;
+    type: PropsType;
 }
 
 export type { ICanvasPropsStyle, PropsType };
