+++ packages/types/dashboard.ts
@@ -1,0 +1,13 @@
+interface IDashboardProps {
+
+}
+
+
+interface IDashboardSidebar {
+
+}
+
+export type {
+    IDashboardSidebar,
+    IDashboardProps
+}
