+++ packages/types/tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "extends": "@repo/typescript-config/base.json",
+    "exclude": ["node_modules"],
+    "compilerOptions": {
+        "esModuleInterop": true,
+        "forceConsistentCasingInFileNames": true,
+        "isolatedModules": true,
+        "moduleResolution": "node",
+        "preserveWatchOutput": true,
+        "skipLibCheck": true,
+        "noEmit": true,
+        "strict": true
+    }
+}
